LPS-67514 Check for errors and use portlet request
When trying to get the blogs entry, and the URL title is provided, check
if there are any errors in the session. This means that the user was
trying to add a new blogs entry with a custom URL title, and some
validation error occurred.

Also, always use the portlet request, as otherwise SessionErrors will
return the wrong value when checking if it is empty.